fix(laravel): defer "filters" dependent services
