+++ src/Laravel/ApiPlatformDeferredProvider.php
@@ -13,8 +13,16 @@ declare(strict_types=1);
 
 namespace ApiPlatform\Laravel;
 
+use ApiPlatform\GraphQl\Resolver\Factory\ResolverFactoryInterface;
+use ApiPlatform\GraphQl\State\Provider\DenormalizeProvider as GraphQlDenormalizeProvider;
+use ApiPlatform\GraphQl\Type\ContextAwareTypeBuilderInterface;
+use ApiPlatform\GraphQl\Type\FieldsBuilder;
+use ApiPlatform\GraphQl\Type\FieldsBuilderEnumInterface;
+use ApiPlatform\GraphQl\Type\TypeConverterInterface;
+use ApiPlatform\GraphQl\Type\TypesContainerInterface;
 use ApiPlatform\JsonApi\Filter\SparseFieldset;
 use ApiPlatform\JsonApi\Filter\SparseFieldsetParameterProvider;
+use ApiPlatform\Laravel\Controller\ApiPlatformController;
 use ApiPlatform\Laravel\Eloquent\Extension\FilterQueryExtension;
 use ApiPlatform\Laravel\Eloquent\Extension\QueryExtensionInterface;
 use ApiPlatform\Laravel\Eloquent\Filter\BooleanFilter;
@@ -26,16 +34,39 @@ use ApiPlatform\Laravel\Eloquent\Filter\JsonApi\SortFilterParameterProvider;
 use ApiPlatform\Laravel\Eloquent\Filter\OrderFilter;
 use ApiPlatform\Laravel\Eloquent\Filter\PartialSearchFilter;
 use ApiPlatform\Laravel\Eloquent\Filter\RangeFilter;
+use ApiPlatform\Laravel\Eloquent\Metadata\Factory\Resource\EloquentResourceCollectionMetadataFactory;
 use ApiPlatform\Laravel\Eloquent\State\CollectionProvider;
 use ApiPlatform\Laravel\Eloquent\State\ItemProvider;
 use ApiPlatform\Laravel\Eloquent\State\LinksHandler;
 use ApiPlatform\Laravel\Eloquent\State\LinksHandlerInterface;
 use ApiPlatform\Laravel\Eloquent\State\PersistProcessor;
 use ApiPlatform\Laravel\Eloquent\State\RemoveProcessor;
+use ApiPlatform\Laravel\Exception\ErrorHandler;
+use ApiPlatform\Laravel\Metadata\CacheResourceCollectionMetadataFactory;
+use ApiPlatform\Laravel\Metadata\ParameterValidationResourceMetadataCollectionFactory;
 use ApiPlatform\Laravel\State\ParameterValidatorProvider;
 use ApiPlatform\Laravel\State\SwaggerUiProcessor;
+use ApiPlatform\Metadata\IdentifiersExtractorInterface;
+use ApiPlatform\Metadata\InflectorInterface;
+use ApiPlatform\Metadata\Operation\PathSegmentNameGeneratorInterface;
+use ApiPlatform\Metadata\Property\Factory\PropertyMetadataFactoryInterface;
+use ApiPlatform\Metadata\Property\Factory\PropertyNameCollectionFactoryInterface;
+use ApiPlatform\Metadata\Resource\Factory\AlternateUriResourceMetadataCollectionFactory;
+use ApiPlatform\Metadata\Resource\Factory\AttributesResourceMetadataCollectionFactory;
+use ApiPlatform\Metadata\Resource\Factory\ConcernsResourceMetadataCollectionFactory;
+use ApiPlatform\Metadata\Resource\Factory\FiltersResourceMetadataCollectionFactory;
+use ApiPlatform\Metadata\Resource\Factory\FormatsResourceMetadataCollectionFactory;
+use ApiPlatform\Metadata\Resource\Factory\InputOutputResourceMetadataCollectionFactory;
+use ApiPlatform\Metadata\Resource\Factory\LinkFactoryInterface;
+use ApiPlatform\Metadata\Resource\Factory\LinkResourceMetadataCollectionFactory;
+use ApiPlatform\Metadata\Resource\Factory\NotExposedOperationResourceMetadataCollectionFactory;
+use ApiPlatform\Metadata\Resource\Factory\OperationNameResourceMetadataCollectionFactory;
+use ApiPlatform\Metadata\Resource\Factory\ParameterResourceMetadataCollectionFactory;
+use ApiPlatform\Metadata\Resource\Factory\PhpDocResourceMetadataCollectionFactory;
 use ApiPlatform\Metadata\Resource\Factory\ResourceMetadataCollectionFactoryInterface;
+use ApiPlatform\Metadata\Resource\Factory\UriTemplateResourceMetadataCollectionFactory;
 use ApiPlatform\Metadata\ResourceAccessCheckerInterface;
+use ApiPlatform\Metadata\ResourceClassResolverInterface;
 use ApiPlatform\Metadata\Util\ReflectionClassRecursiveIterator;
 use ApiPlatform\Serializer\Filter\FilterInterface as SerializerFilterInterface;
 use ApiPlatform\Serializer\Filter\PropertyFilter;
@@ -50,9 +81,14 @@ use ApiPlatform\State\Provider\DeserializeProvider;
 use ApiPlatform\State\Provider\ParameterProvider;
 use ApiPlatform\State\Provider\SecurityParameterProvider;
 use ApiPlatform\State\ProviderInterface;
+use Illuminate\Contracts\Debug\ExceptionHandler as ExceptionHandlerInterface;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\Support\DeferrableProvider;
 use Illuminate\Support\ServiceProvider;
+use Negotiation\Negotiator;
+use Psr\Log\LoggerInterface;
+use Symfony\Component\Serializer\NameConverter\CamelCaseToSnakeCaseNameConverter;
+use Symfony\Component\Serializer\NameConverter\NameConverterInterface;
 
 class ApiPlatformDeferredProvider extends ServiceProvider implements DeferrableProvider
 {
@@ -149,6 +185,130 @@ class ApiPlatformDeferredProvider extends ServiceProvider implements DeferrableP
         });
 
         $this->autoconfigure($classes, ProviderInterface::class, [ItemProvider::class, CollectionProvider::class, ErrorProvider::class]);
+
+        $this->app->singleton(ResourceMetadataCollectionFactoryInterface::class, function (Application $app) {
+            /** @var ConfigRepository $config */
+            $config = $app['config'];
+            $formats = $config->get('api-platform.formats');
+
+            if ($config->get('api-platform.swagger_ui.enabled', false) && !isset($formats['html'])) {
+                $formats['html'] = ['text/html'];
+            }
+
+            return new CacheResourceCollectionMetadataFactory(
+                new EloquentResourceCollectionMetadataFactory(
+                    new ParameterValidationResourceMetadataCollectionFactory(
+                        new ParameterResourceMetadataCollectionFactory(
+                            $this->app->make(PropertyNameCollectionFactoryInterface::class),
+                            $this->app->make(PropertyMetadataFactoryInterface::class),
+                            new AlternateUriResourceMetadataCollectionFactory(
+                                new FiltersResourceMetadataCollectionFactory(
+                                    new FormatsResourceMetadataCollectionFactory(
+                                        new InputOutputResourceMetadataCollectionFactory(
+                                            new PhpDocResourceMetadataCollectionFactory(
+                                                new OperationNameResourceMetadataCollectionFactory(
+                                                    new LinkResourceMetadataCollectionFactory(
+                                                        $app->make(LinkFactoryInterface::class),
+                                                        new UriTemplateResourceMetadataCollectionFactory(
+                                                            $app->make(LinkFactoryInterface::class),
+                                                            $app->make(PathSegmentNameGeneratorInterface::class),
+                                                            new NotExposedOperationResourceMetadataCollectionFactory(
+                                                                $app->make(LinkFactoryInterface::class),
+                                                                new AttributesResourceMetadataCollectionFactory(
+                                                                    new ConcernsResourceMetadataCollectionFactory(
+                                                                        null,
+                                                                        $app->make(LoggerInterface::class),
+                                                                        $config->get('api-platform.defaults', []),
+                                                                        $config->get('api-platform.graphql.enabled'),
+                                                                    ),
+                                                                    $app->make(LoggerInterface::class),
+                                                                    $config->get('api-platform.defaults', []),
+                                                                    $config->get('api-platform.graphql.enabled'),
+                                                                ),
+                                                            )
+                                                        ),
+                                                        $config->get('api-platform.graphql.enabled')
+                                                    )
+                                                )
+                                            )
+                                        ),
+                                        $formats,
+                                        $config->get('api-platform.patch_formats'),
+                                    )
+                                )
+                            ),
+                            $app->make('filters'),
+                            $app->make(CamelCaseToSnakeCaseNameConverter::class),
+                            $this->app->make(LoggerInterface::class)
+                        ),
+                        $app->make('filters')
+                    )
+                ),
+                true === $config->get('app.debug') ? 'array' : $config->get('api-platform.cache', 'file')
+            );
+        });
+
+        $this->app->singleton(
+            ExceptionHandlerInterface::class,
+            function (Application $app) {
+                /** @var ConfigRepository */
+                $config = $app['config'];
+
+                return new ErrorHandler(
+                    $app,
+                    $app->make(ResourceMetadataCollectionFactoryInterface::class),
+                    $app->make(ApiPlatformController::class),
+                    $app->make(IdentifiersExtractorInterface::class),
+                    $app->make(ResourceClassResolverInterface::class),
+                    $app->make(Negotiator::class),
+                    $config->get('api-platform.exception_to_status'),
+                    $config->get('app.debug')
+                );
+            }
+        );
+
+        if (interface_exists(FieldsBuilderEnumInterface::class)) {
+            $this->registerGraphQl();
+        }
+    }
+
+    private function registerGraphQl(): void
+    {
+        $this->app->singleton('api_platform.graphql.state_provider.parameter', function (Application $app) {
+            $tagged = iterator_to_array($app->tagged(ParameterProviderInterface::class));
+            $tagged['api_platform.serializer.filter_parameter_provider'] = $app->make(SerializerFilterParameterProvider::class);
+
+            return new ParameterProvider(
+                new ParameterValidatorProvider(
+                    new SecurityParameterProvider(
+                        $app->make(GraphQlDenormalizeProvider::class),
+                        $app->make(ResourceAccessCheckerInterface::class)
+                    ),
+                ),
+                new ServiceLocator($tagged)
+            );
+        });
+
+        $this->app->singleton(FieldsBuilderEnumInterface::class, function (Application $app) {
+            /** @var ConfigRepository */
+            $config = $app['config'];
+
+            return new FieldsBuilder(
+                $app->make(PropertyNameCollectionFactoryInterface::class),
+                $app->make(PropertyMetadataFactoryInterface::class),
+                $app->make(ResourceMetadataCollectionFactoryInterface::class),
+                $app->make(ResourceClassResolverInterface::class),
+                $app->make(TypesContainerInterface::class),
+                $app->make(ContextAwareTypeBuilderInterface::class),
+                $app->make(TypeConverterInterface::class),
+                $app->make(ResolverFactoryInterface::class),
+                $app->make('filters'),
+                $app->make(Pagination::class),
+                $app->make(NameConverterInterface::class),
+                $config->get('api-platform.graphql.nesting_separator') ?? '__',
+                $app->make(InflectorInterface::class)
+            );
+        });
     }
 
     /**
@@ -186,6 +346,9 @@ class ApiPlatformDeferredProvider extends ServiceProvider implements DeferrableP
             ParameterProvider::class,
             FilterQueryExtension::class,
             'filters',
+            ResourceMetadataCollectionFactoryInterface::class,
+            'api_platform.graphql.state_provider.parameter',
+            FieldsBuilderEnumInterface::class,
         ];
     }
 }
